Drop support for SQL Anywhere <16
