+++ lib/Doctrine/DBAL/Platforms/Keywords/SQLAnywhereKeywords.php
@@ -28,6 +28,7 @@ class SQLAnywhereKeywords extends KeywordList
             'ALTER',
             'AND',
             'ANY',
+            'ARRAY',
             'AS',
             'ASC',
             'ATTACH',
@@ -68,6 +69,7 @@ class SQLAnywhereKeywords extends KeywordList
             'CURRENT_USER',
             'CURSOR',
             'DATE',
+            'DATETIMEOFFSET',
             'DBSPACE',
             'DEALLOCATE',
             'DEC',
@@ -114,7 +116,6 @@ class SQLAnywhereKeywords extends KeywordList
             'IF',
             'IN',
             'INDEX',
-            'INDEX_LPAREN',
             'INNER',
             'INOUT',
             'INSENSITIVE',
@@ -127,20 +128,22 @@ class SQLAnywhereKeywords extends KeywordList
             'INTEGRATED',
             'INTERSECT',
             'INTO',
-            'IQ',
             'IS',
             'ISOLATION',
             'JOIN',
+            'JSON',
             'KERBEROS',
             'KEY',
             'LATERAL',
             'LEFT',
             'LIKE',
+            'LIMIT',
             'LOCK',
             'LOGIN',
             'LONG',
             'MATCH',
             'MEMBERSHIP',
+            'MERGE',
             'MESSAGE',
             'MODE',
             'MODIFY',
@@ -158,6 +161,8 @@ class SQLAnywhereKeywords extends KeywordList
             'OFF',
             'ON',
             'OPEN',
+            'OPENSTRING',
+            'OPENXML',
             'OPTION',
             'OPTIONS',
             'OR',
@@ -194,6 +199,8 @@ class SQLAnywhereKeywords extends KeywordList
             'RIGHT',
             'ROLLBACK',
             'ROLLUP',
+            'ROW',
+            'ROWTYPE',
             'SAVE',
             'SAVEPOINT',
             'SCROLL',
@@ -205,6 +212,7 @@ class SQLAnywhereKeywords extends KeywordList
             'SHARE',
             'SMALLINT',
             'SOME',
+            'SPATIAL',
             'SQLCODE',
             'SQLSTATE',
             'START',
@@ -212,7 +220,6 @@ class SQLAnywhereKeywords extends KeywordList
             'SUBTRANS',
             'SUBTRANSACTION',
             'SYNCHRONIZE',
-            'SYNTAX_ERROR',
             'TABLE',
             'TEMPORARY',
             'THEN',
@@ -222,6 +229,7 @@ class SQLAnywhereKeywords extends KeywordList
             'TO',
             'TOP',
             'TRAN',
+            'TREAT',
             'TRIGGER',
             'TRUNCATE',
             'TSEQUAL',
@@ -230,6 +238,7 @@ class SQLAnywhereKeywords extends KeywordList
             'UNIQUE',
             'UNIQUEIDENTIFIER',
             'UNKNOWN',
+            'UNNEST',
             'UNSIGNED',
             'UPDATE',
             'UPDATING',
@@ -241,6 +250,7 @@ class SQLAnywhereKeywords extends KeywordList
             'VARBIT',
             'VARCHAR',
             'VARIABLE',
+            'VARRAY',
             'VARYING',
             'VIEW',
             'WAIT',
@@ -250,9 +260,6 @@ class SQLAnywhereKeywords extends KeywordList
             'WHILE',
             'WINDOW',
             'WITH',
-            'WITH_CUBE',
-            'WITH_LPAREN',
-            'WITH_ROLLUP',
             'WITHIN',
             'WORK',
             'WRITETEXT',
