+++ tests/Doctrine/Tests/DBAL/Driver/AbstractSQLAnywhereDriverTest.php
@@ -4,9 +4,6 @@ namespace Doctrine\Tests\DBAL\Driver;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Driver\AbstractSQLAnywhereDriver;
-use Doctrine\DBAL\Platforms\SQLAnywhere11Platform;
-use Doctrine\DBAL\Platforms\SQLAnywhere12Platform;
-use Doctrine\DBAL\Platforms\SQLAnywhere16Platform;
 use Doctrine\DBAL\Platforms\SQLAnywherePlatform;
 use Doctrine\DBAL\Schema\SQLAnywhereSchemaManager;
 
@@ -19,7 +16,7 @@ class AbstractSQLAnywhereDriverTest extends AbstractDriverTest
 
     protected function createPlatform()
     {
-        return new SQLAnywhere12Platform();
+        return new SQLAnywherePlatform();
     }
 
     protected function createSchemaManager(Connection $connection)
@@ -30,35 +27,13 @@ class AbstractSQLAnywhereDriverTest extends AbstractDriverTest
     protected function getDatabasePlatformsForVersions()
     {
         return [
-            ['10', SQLAnywherePlatform::class],
-            ['10.0', SQLAnywherePlatform::class],
-            ['10.0.0', SQLAnywherePlatform::class],
-            ['10.0.0.0', SQLAnywherePlatform::class],
-            ['10.1.2.3', SQLAnywherePlatform::class],
-            ['10.9.9.9', SQLAnywherePlatform::class],
-            ['11', SQLAnywhere11Platform::class],
-            ['11.0', SQLAnywhere11Platform::class],
-            ['11.0.0', SQLAnywhere11Platform::class],
-            ['11.0.0.0', SQLAnywhere11Platform::class],
-            ['11.1.2.3', SQLAnywhere11Platform::class],
-            ['11.9.9.9', SQLAnywhere11Platform::class],
-            ['12', SQLAnywhere12Platform::class],
-            ['12.0', SQLAnywhere12Platform::class],
-            ['12.0.0', SQLAnywhere12Platform::class],
-            ['12.0.0.0', SQLAnywhere12Platform::class],
-            ['12.1.2.3', SQLAnywhere12Platform::class],
-            ['12.9.9.9', SQLAnywhere12Platform::class],
-            ['13', SQLAnywhere12Platform::class],
-            ['14', SQLAnywhere12Platform::class],
-            ['15', SQLAnywhere12Platform::class],
-            ['15.9.9.9', SQLAnywhere12Platform::class],
-            ['16', SQLAnywhere16Platform::class],
-            ['16.0', SQLAnywhere16Platform::class],
-            ['16.0.0', SQLAnywhere16Platform::class],
-            ['16.0.0.0', SQLAnywhere16Platform::class],
-            ['16.1.2.3', SQLAnywhere16Platform::class],
-            ['16.9.9.9', SQLAnywhere16Platform::class],
-            ['17', SQLAnywhere16Platform::class],
+            ['16', SQLAnywherePlatform::class],
+            ['16.0', SQLAnywherePlatform::class],
+            ['16.0.0', SQLAnywherePlatform::class],
+            ['16.0.0.0', SQLAnywherePlatform::class],
+            ['16.1.2.3', SQLAnywherePlatform::class],
+            ['16.9.9.9', SQLAnywherePlatform::class],
+            ['17', SQLAnywherePlatform::class],
         ];
     }
