+++ tests/Doctrine/Tests/DBAL/Platforms/SQLAnywherePlatformTest.php
@@ -13,6 +13,7 @@ use Doctrine\DBAL\Schema\Comparator;
 use Doctrine\DBAL\Schema\Constraint;
 use Doctrine\DBAL\Schema\ForeignKeyConstraint;
 use Doctrine\DBAL\Schema\Index;
+use Doctrine\DBAL\Schema\Sequence;
 use Doctrine\DBAL\Schema\Table;
 use Doctrine\DBAL\Schema\TableDiff;
 use Doctrine\DBAL\Schema\UniqueConstraint;
@@ -453,17 +454,120 @@ class SQLAnywherePlatformTest extends AbstractPlatformTestCase
     public function testGeneratesCreateIndexWithAdvancedPlatformOptionsSQL()
     {
         self::assertEquals(
-            'CREATE VIRTUAL UNIQUE CLUSTERED INDEX fooindex ON footable (a, b) FOR OLAP WORKLOAD',
+            'CREATE UNIQUE INDEX fooindex ON footable (a, b) WITH NULLS DISTINCT',
             $this->platform->getCreateIndexSQL(
                 new Index(
                     'fooindex',
                     ['a', 'b'],
                     true,
                     false,
-                    ['virtual', 'clustered', 'for_olap_workload']
+                    ['with_nulls_distinct']
                 ),
                 'footable'
             )
+        );
+
+        // WITH NULLS DISTINCT clause not available on primary indexes.
+        self::assertEquals(
+            'ALTER TABLE footable ADD PRIMARY KEY (a, b)',
+            $this->platform->getCreateIndexSQL(
+                new Index(
+                    'fooindex',
+                    ['a', 'b'],
+                    false,
+                    true,
+                    ['with_nulls_distinct']
+                ),
+                'footable'
+            )
+        );
+
+        // WITH NULLS DISTINCT clause not available on non-unique indexes.
+        self::assertEquals(
+            'CREATE INDEX fooindex ON footable (a, b)',
+            $this->platform->getCreateIndexSQL(
+                new Index(
+                    'fooindex',
+                    ['a', 'b'],
+                    false,
+                    false,
+                    ['with_nulls_distinct']
+                ),
+                'footable'
+            )
+        );
+
+        self::assertEquals(
+            'CREATE VIRTUAL UNIQUE CLUSTERED INDEX fooindex ON footable (a, b) WITH NULLS NOT DISTINCT FOR OLAP WORKLOAD',
+            $this->platform->getCreateIndexSQL(
+                new Index(
+                    'fooindex',
+                    ['a', 'b'],
+                    true,
+                    false,
+                    ['virtual', 'clustered', 'with_nulls_not_distinct', 'for_olap_workload']
+                ),
+                'footable'
+            )
+        );
+        self::assertEquals(
+            'CREATE VIRTUAL CLUSTERED INDEX fooindex ON footable (a, b) FOR OLAP WORKLOAD',
+            $this->platform->getCreateIndexSQL(
+                new Index(
+                    'fooindex',
+                    ['a', 'b'],
+                    false,
+                    false,
+                    ['virtual', 'clustered', 'with_nulls_not_distinct', 'for_olap_workload']
+                ),
+                'footable'
+            )
+        );
+
+        // WITH NULLS NOT DISTINCT clause not available on primary indexes.
+        self::assertEquals(
+            'ALTER TABLE footable ADD PRIMARY KEY (a, b)',
+            $this->platform->getCreateIndexSQL(
+                new Index(
+                    'fooindex',
+                    ['a', 'b'],
+                    false,
+                    true,
+                    ['with_nulls_not_distinct']
+                ),
+                'footable'
+            )
+        );
+
+        // WITH NULLS NOT DISTINCT clause not available on non-unique indexes.
+        self::assertEquals(
+            'CREATE INDEX fooindex ON footable (a, b)',
+            $this->platform->getCreateIndexSQL(
+                new Index(
+                    'fooindex',
+                    ['a', 'b'],
+                    false,
+                    false,
+                    ['with_nulls_not_distinct']
+                ),
+                'footable'
+            )
+        );
+    }
+
+    public function testThrowsExceptionOnInvalidWithNullsNotDistinctIndexOptions()
+    {
+        $this->expectException('UnexpectedValueException');
+
+        $this->platform->getCreateIndexSQL(
+            new Index(
+                'fooindex',
+                ['a', 'b'],
+                false,
+                false,
+                ['with_nulls_distinct', 'with_nulls_not_distinct']
+            ),
+            'footable'
         );
     }
 
@@ -574,19 +678,28 @@ class SQLAnywherePlatformTest extends AbstractPlatformTestCase
         );
     }
 
-    public function testDoesNotSupportRegexp()
-    {
-        $this->expectException(DBALException::class);
-
-        $this->platform->getRegexpExpression();
-    }
-
     public function testHasCorrectDateTimeTzFormatString()
     {
-        // Date time type with timezone is not supported before version 12.
-        // For versions before we have to ensure that the date time with timezone format
-        // equals the normal date time format so that it corresponds to the declaration SQL equality (datetimetz -> datetime).
-        self::assertEquals($this->platform->getDateTimeFormatString(), $this->platform->getDateTimeTzFormatString());
+        self::assertEquals('Y-m-d H:i:s.uP', $this->platform->getDateTimeTzFormatString());
+    }
+
+    public function testGeneratesDateTimeTzColumnTypeDeclarationSQL()
+    {
+        self::assertEquals(
+            'TIMESTAMP WITH TIME ZONE',
+            $this->platform->getDateTimeTzTypeDeclarationSQL([
+                'length' => 10,
+                'fixed' => true,
+                'unsigned' => true,
+                'autoincrement' => true,
+            ])
+        );
+    }
+
+    public function testInitializesDateTimeTzTypeMapping()
+    {
+        self::assertTrue($this->platform->hasDoctrineTypeMappingFor('timestamp with time zone'));
+        self::assertEquals('datetime', $this->platform->getDoctrineTypeMapping('timestamp with time zone'));
     }
 
     public function testHasCorrectDefaultTransactionIsolationLevel()
@@ -745,7 +858,41 @@ class SQLAnywherePlatformTest extends AbstractPlatformTestCase
 
     public function testDoesNotSupportSequences()
     {
-        self::assertFalse($this->platform->supportsSequences());
+        self::markTestSkipped('This version of the platform now supports sequences.');
+    }
+
+    public function testSupportsSequences()
+    {
+        self::assertTrue($this->platform->supportsSequences());
+    }
+
+    public function testGeneratesSequenceSqlCommands()
+    {
+        $sequence = new Sequence('myseq', 20, 1);
+        self::assertEquals(
+            'CREATE SEQUENCE myseq INCREMENT BY 20 START WITH 1 MINVALUE 1',
+            $this->platform->getCreateSequenceSQL($sequence)
+        );
+        self::assertEquals(
+            'ALTER SEQUENCE myseq INCREMENT BY 20',
+            $this->platform->getAlterSequenceSQL($sequence)
+        );
+        self::assertEquals(
+            'DROP SEQUENCE myseq',
+            $this->platform->getDropSequenceSQL('myseq')
+        );
+        self::assertEquals(
+            'DROP SEQUENCE myseq',
+            $this->platform->getDropSequenceSQL($sequence)
+        );
+        self::assertEquals(
+            'SELECT myseq.NEXTVAL',
+            $this->platform->getSequenceNextValSQL('myseq')
+        );
+        self::assertEquals(
+            'SELECT sequence_name, increment_by, start_with, min_value FROM SYS.SYSSEQUENCE',
+            $this->platform->getListSequencesSQL(null)
+        );
     }
 
     public function testDoesNotSupportInlineColumnComments()
